API

Simulation Information Type and Methods

VariationalMC.SimulationInfo — Type

SimulationInfo

Contains identification information about simulation, including the location data is written to, the simulation ID, and MPI process ID, and whether this simulation started a new simulation or resumed a previous simulation.

Fields

filepath::String: File path to where data folder lives.

datafolder_prefix: Prefix for the data folder name.

datafolder_name::String: The data folder name, given by $(datafolder_prefix)_$(sID).

datafolder::String: The data folder, including filepath, given by joinpath(filepath, datafolder_name).

pID::Int: MPI process ID, defaults to 0 if MPI not being used.

sID::Int: Simulation ID.

bin_files::Vector{Vector{UInt8}}: Represents the HDF5 files containing the binned data.

resuming::Bool: Whether current simulation is resuming a previous simulation (true) or starting a new one (false).

variationalmc_version::VersionNumber: Version of VariationalMC.jl used in simulation.

VariationalMC.SimulationInfo — Method

SimulationInfo(;
    # KEYWORD ARGUMENTS
    datafolder_prefix::String,
    filepath::String = ".",
    sID::Int=0,
    pID::Int=0
)

Initialize and return in instance of the type SimulationInfo.

VariationalMC.initialize_datafolder — Function

initialize_datafolder(comm::MPI.Comm, sim_info::SimulationInfo)

initialize_datafolder(sim_info::SimulationInfo)

Initialize sim_info.datafolder directory if it does not already exist. If comm::MPI.Comm is passed as the first argument, this this function will synchronize all the MPI processes, ensuring that none proceed beyond this function call until the data folder that results will be written to is successfully initialized.

VariationalMC.model_summary — Function

model_summary(;
    simulation_info::SimulationInfo,
    model_geometry::ModelGeometry,
    particle_configuration::ParticleConfiguration,
    tight_binding_model::TightBindingModel,
    interactions::Union{Tuple,Nothing} = nothing,
    parameters::Union{Tuple, Nothing} = nothing
)

Write model to summary to file. This includes information on the canonical ensemble configuration, tight binding model, interactions, and variational parameters.

Model Geometry Type and Methods

VariationalMC.ModelGeometry — Type

ModelGeometry{D, T<:AbstractFloat, N}

Contains all the information defining the lattice geometry for the model in D spatial dimensions.

Comment

The bond ID associated with a bond::Bond{D} corresponds to the index associated with it into the bonds vector field.

Fields

unit_cell::UnitCell{D,T,N}: Defines unit cell.

lattice::Lattice{D}: Defines finite lattice extent.

bonds::Vector{Bond{D}}: All available bond definitions in simulation, with vector indices giving the bond ID.

VariationalMC.ModelGeometry — Method

ModelGeometry(  unit_cell::UnitCell, 
                lattice::Lattice ) where {D}

Initialize and return a ModelGeometry instance. Defines a "trivial" bond definition for each orbital in the unit cell that connects an orbital to itself.

VariationalMC.add_bond! — Function

add_bond!(  model_geometry::ModelGeometry{D,T}, 
            bond::Bond{D}) where {D, T}

Add bond definition to model_geometry, returning the bond ID i.e. the index to bond in the vector model_geometry.bonds. This method first checks that bond is not already defined. If it is this method simply returns the corresponding bond ID. If bond is not already defined, then it is appended to the vector model_geometry.bonds.

VariationalMC.get_bond_id — Function

get_bond_id(model_geometry::ModelGeometry{D,T}, bond::Bond{D}) where {D, T}

Return the bond ID associated with the bond defintion bond, returning bond_id=0 if the it is not a recorded bond.

Particle Configuration Type and Methods

VariationalMC.ParticleConfiguration — Type

ParticleConfiguration( 
    density::AbstractFloat; 
    # KEYWORD ARGUMENTS
    model_geometry::ModelGeometry,
    particle_hole_transform::Bool 
) where {E<:AbstractFloat}

Given an initial particle density, initialize and return an instance of the type ParticleConfiguration.

density::AbstractFloat: desired electronic density.

KEYWORD ARGUMENTS

model_geometry::ModelGeometry: contains unit cell and lattice quantities.

ph_transform::Bool: Whether configuration will be particle-hole transformed.

ParticleConfiguration( 
    Np::Int;
    # KEYWORD ARGUMENTS
    model_geometry::ModelGeometry,
    particle_hole_transform::Bool 
) where {I<:Integer}

Given an initial particle number, initialize and return an instance of the type ParticleConfiguration.

Np::Int: total number of particles.

KEYWORD ARGUMENTS

model_geometry::ModelGeometry: contains unit cell and lattice quantities.

ph_transform::Bool: Whether configuration is particle-hole transformed.

ParticleConfiguration(;
    # KEYWORD ARGUMENTS
    nup::Int, 
    ndn::Int,
    model_geometry::ModelGeometry, 
    particle_hole_transform::Bool 
) where {I<:Integer}

Given a certain number of spin-up and spin-down particles, initialize and return an instance of the type ParticleConfiguration.

KEYWORD ARGUMENTS

nup::Int: number of spin-up fermions.

ndn::Int: number of spin-down fermions.

model_geometry::ModelGeometry: contains unit cell and lattice quantities.

ph_transform::Bool: Whether configuration is particle-hole transformed.

VariationalMC.ParticleConfiguration — Method

ParticleConfiguration(;
    # KEYWORD ARGUMENTS
    nup::Int, 
    ndn::Int,
    model_geometry::ModelGeometry, 
    particle_hole_transform::Bool 
) where {I<:Integer}

Given a certain number of spin-up and spin-down particles, initialize and return an instance of the type ParticleConfiguration.

KEYWORD ARGUMENTS

nup::Int: number of spin-up fermions.

ndn::Int: number of spin-down fermions.

model_geometry::ModelGeometry: contains unit cell and lattice quantities.

ph_transform::Bool: Whether configuration is particle-hole transformed.

Tight-Binding Model and Parameters

VariationalMC.TightBindingModel — Type

TightBindingModel{T<:Number, E<:AbstractFloat, D}

Defines a non-interacting tight binding model in D dimensions.

Fields

μ::E: Initial value of the chemical potential.

t_bond_ids::Vector{Int}: The bond ID for each bond/hopping definition.

t_bonds::Vector{Bond{D}}: Bond definition for each type of hopping in the tight binding model.

t_mean::Vector{T}: Mean hopping energy for each type of hopping.

η::Vector{E}: Relative twist angle for each lattice vector direction $d \in [1, D]$ such that $\eta_d \in [0, 1).$

expniϕ::Vector{T}: Twist angle phase $\exp(i n_d \phi_d)$ for each hopping.

VariationalMC.TightBindingModel — Method

TightBindingModel(;
    # KEYWORD ARGUMENTS
    model_geometry::ModelGeometry{D,E,N},
    μ::E,
    t_bonds::Vector{Bond{D}} = Bond{ndims(model_geometry.unit_cell)}[],
    t_mean::Vector{T} = T[],
    η::Union{Vector{E},Nothing} = nothing
) where {T<:Number, E<:AbstractFloat, D, N}

Initialize and return an instance of TightBindingModel, also adding/recording the bond defintions t_bonds to the ModelGeometry instance model_geometry.

VariationalMC.TightBindingParameters — Type

TightBindingParameters{T<:Number, E<:AbstractFloat}

A mutable struct containing all the parameters needed to characterize a finite tight-binding Hamiltonian for a single spin species $\sigma$ on a finite lattice with periodic boundary conditions of the form

\[\hat{H}_{0,\sigma}=-\sum_{\langle i,j\rangle}(t_{ij} \hat{c}_{\sigma,i}^{\dagger}\hat{c}_{\sigma,j}+\textrm{h.c.}),\]

where $\hat{c}_{\sigma,i}^\dagger$ is the fermion creation operator for an electron with spin $\sigma$ on orbital $i,$ $t_{i,j}$ are the hopping energies, and $\mu$ is the chemical potential.

Fields

μ::E: The chemical potential $\mu.$

const t::Vector{T}: The hopping energy $t_{i,j}$ associated with each pair of neighboring orbitals connected by a bond in the lattice.

const neighbor_table::Matrix{Int}: Neighbor table containing all pairs of orbitals in the lattices connected by a bond, with a non-zero hopping energy between them.

const neighbor_map::Dict: Dictionary containing all sites in the lattice, mapping them to their valid neighboring sites and the bonds between them.

const bond_ids::Vector{Int}: The bond ID definitions that define the types of hopping in the lattice.

const bond_slices::Vector{UnitRange{Int}}: Slices of neighbor_table corresponding to given bond ID i.e. the neighbors neighbor_table[:,bond_slices[i]] corresponds the bond_ids[i] bond defintion.

const norbital::Int: Number of orbitals per unit cell.

VariationalMC.TightBindingParameters — Method

TightBindingParameters(;
    # KEYWORD ARGUMENTS
    tight_binding_model::TightBindingModel{T,E,D},
    particle_configuration::ParticleConfiguration{E,I},
    model_geometry::ModelGeometry{D,E}
) where {T,E,D}

Initialize and return an instance of TightBindingParameters.

VariationalMC.measure_hopping_energy — Function

measure_hopping_energy(
    detwf::DeterminantalWavefunction{T},
    tight_binding_parameters::TightBindingParameters{T,E},
    pconfig::Vector{Int},
    hopping_id::Int,
    N::Int,
    ph_transform::Bool;
    # KEYWORD ARGUMENTS
    model_geometry::ModelGeometry = model_geometry,
    jas_factors::Union{Tuple{<:AbstractJastrowFactor{T}},Nothing} = nothing,
    jas_parameters::Union{Tuple{JastrowParameters{T}},Nothing} = nothing
) where {T<:Number, E<:AbstractFloat}

Calculate the average hopping energy $-\langle t_{\langle i,j \rangle} \hat{c}^\dagger_{\sigma,i} \hat{c}_{\sigma,j} + {\rm h.c.} \rangle$.

Hubbard Model and Parameters

VariationalMC.HubbardModel — Type

HubbardModel{T<:AbstractFloat}

A type to represent a multi-orbital Hubbard model; specifically, the particle-hole asymmetric form for the Hubbard interaction

\[\hat{H}_{U}=\sum_{\mathbf{i},\nu}U_{\nu,\mathbf{i}}\hat{n}_{\uparrow,\nu,\mathbf{i}}\hat{n}_{\downarrow,\nu,\mathbf{i}},\]

where $\mathbf{i}$ specifies the unit cell, and $\nu$ denotes the orbital in the unit cell.

Fields

U_orbital_ids::Vector{Int}: Orbital species/IDs in unit cell with finite Hubbard interaction.

U_mean::Vector{T}: Average Hubbard interaction strength $U_\nu$ for a given orbital species in the lattice.

VariationalMC.HubbardModel — Method

HubbardModel(;
    # KEYWORD ARGUMENTS
    U_orbital::AbstractVector{Int},
    U_mean::AbstractVector{T}
) where {T<:AbstractFloat}

Initialize and return an instance of the type HubbardModel.

Keyword Arguments

U_orbital::Vector{Int}: Orbital species/IDs in unit cell with finite Hubbard interaction.

U_mean::Vector{T}: Average Hubbard interaction strength $U_\nu$ for a given orbital species in the lattice.

VariationalMC.HubbardParameters — Type

HubbardParameters{T<:AbstractFloat}

Hubbard parameters for finite lattice.

Fields

U::Vector{T}: On-site Hubbard interaction for each site with finite Hubbard interaction.

sites::Vector{Int}: Site index associated with each finite Hubbard U interaction.

orbital_ids::Vector{Int}: Orbital ID/species in unit cell with finite Hubbard interaction.

VariationalMC.HubbardParameters — Method

HubbardParameters(;
    hubbard_model::HubbardModel{T},
    model_geometry::ModelGeometry{D,T}
) where {D, T<:AbstractFloat}

Initialize an instance of HubbardParameters.

VariationalMC.ExtendedHubbardModel — Type

ExtendedHubbardModel{T<:AbstractFloat}

A type to represent extended Hubbard interactions; specifically, the particle-hole asymmetric form of the extended Hubbard interaction

\[\begin{align*} \hat{H}_{V} = \sum_{\mathbf{j},\mathbf{r},\nu,\eta}V_{(\mathbf{j}+\mathbf{r},\nu),(\mathbf{j},\eta)} & \hat{n}_{\mathbf{j}+\mathbf{r},\nu}\hat{n}_{\mathbf{j},\eta} \\ = \sum_{\mathbf{j},\mathbf{r},\nu,\eta}V_{(\mathbf{j}+\mathbf{r},\nu),(\mathbf{j},\eta)} & \bigg[\tfrac{1}{2}(\hat{n}_{\mathbf{j}+\mathbf{r},\nu}+\hat{n}_{\mathbf{j},\eta}-2)^{2}-1 \\ & -\hat{n}_{\mathbf{j}+\mathbf{r},\nu,\uparrow}\hat{n}_{\mathbf{j}+\mathbf{r},\nu\downarrow}-\hat{n}_{\mathbf{j},\eta,\uparrow}\hat{n}_{\mathbf{j},\eta\downarrow}+\tfrac{3}{2}\hat{n}_{\mathbf{j}+\mathbf{r},\nu}+\tfrac{3}{2}\hat{n}_{\mathbf{j},\eta}\bigg], \end{align*}\]

where $\mathbf{j}$ specifies a unit cell in the lattice, $\mathbf{r}$ is a displacement in units, and $\nu$ and $\eta$ specify the orbital in a given unit cell. Here, $\hat{n}_{\mathbf{j},\eta} = (\hat{n}_{\uparrow,\mathbf{j},\eta} + \hat{n}_{\downarrow,\mathbf{j},\eta})$ is the electron number operator for orbital $\eta$ in unit cell $\mathbf{j}$ in the lattice. $V_{(\mathbf{j}+\mathbf{r},\nu),(\mathbf{j},\eta)}$ controls the strength of the extended Hubbard interaction between orbital $\eta$ in unit cell $\mathbf{j}$ and orbital $\nu$ in unit cell $\mathbf{j}+\mathbf{r}$.

Fields

V_bond_ids::Vector{Int}: Bond IDs specifying bond definition that separates a pair of orbitals with an extended Hubbard interaction between them.

V_mean::Vector{T}: Average extended Hubbard interaction strength $V_{(\mathbf{j}+\mathbf{r},\nu),(\mathbf{j},\eta)}$ associated with bond definition.

VariationalMC.ExtendedHubbardModel — Method

ExtendedHubbardModel(;
    # KEYWORD ARGUMENTS
    model_geometry::ModelGeometry{D,T},
    V_bond::Vector{Bond{D}},
    V_mean::Vector{T},
) where {T<:AbstractFloat, D}

Initialize and return an instance of the type ExtendedHubbardModel.

VariationalMC.ExtendedHubbardParameters — Type

ExtendedHubbardParameters{T<:AbstractFloat}

Extended Hubbard interaction parameters for finite lattice.

Fields

V::Vector{T}: Extended Hubbard interaction strength for each pair neighbors in the lattice.

neighbor_table::Matrix{Int}: Neighbor table for extended Hubbard interactions on lattice.

bond_ids::Vector{Int}: Bond IDs used to define extended Hubbard interactions.

VariationalMC.ExtendedHubbardParameters — Method

ExtendedHubbardParameters(;
    # KEYWORD ARGUMENTS
    extended_hubbard_model::ExtendedHubbardModel{T},
    model_geometry::ModelGeometry{D,T}
) where {D, T<:AbstractFloat}

Initialize an instance of the ExtendedHubbardParameters type.

VariationalMC.measure_hubbard_energy — Function

measure_hubbard_energy(
    hubbard_parameters::HubbardParameters{E},
    hubbard_id::Int,
    pconfig::Vector{Int},
    ph_transform::Bool
) where {E<:AbstractFloat}

Calculates the average Hubbard energy $U \langle \hat{n}_\uparrow \hat{n}_\downarrow \rangle$.

ARGUMENTS

hubbard_parameters::HubbardParameters{E}: Instance of HubbardParameters.

hubbard_id::Int: ID of the Hubbard interaction.

pconfig::Vector{Int}: Current particle configuration.

ph_transform::Bool: Whether the model is particle-hole transformed.

VariationalMC.measure_ext_hubbard_energy — Function

measure_ext_hubbard_energy(
    extended_hubbard_parameters::ExtendedHubbardParameters{E},
    ext_hubbard_id::Int,
    pconfig::Vector{Int},
    ph_transform::Bool
) where {E<:AbstractFloat}

Calculates the average extended Hubbard interaction energy $V \hat{n}_i \hat{n}_j$.

ARGUMENTS

extended_hubbard_parameters::ExtendedHubbardParameters{E}: Instance of ExtendedHubbardParameters.

ext_hubbard_id::Int: ID of the extended Hubbard interaction.

pconfig::Vector{Int}: Current particle configuration.

ph_transform::Bool: Whether model is particle-hole transformed.

Variational Parameters

Variational Parameter Names

Determinantal Parameters

Jastrow Parameters

Variational Parameter Names

VariationalMC.DETERMINANTAL_PARAMETERS — Constant

DETERMINANTAL_PARAMETERS = Base.ImmutableDict(
    "μ"             => "charge", 
    "density"       => "charge", 
    "spin-x"        => "spin", 
    "spin-z"        => "spin", 
    "s-wave"        => "pair", 
    "d-wave"        => "pair"
)

List of all determinantal parameters that can be added to the wavefunction, along with the corresponding ordering type.

VariationalMC.JASTROW_PARAMETERS — Constant

JASTROW_PARAMETERS = Base.ImmutableDict(
    "electron-electron" => ["density-density", "spin-spin"],
    "phonon-phonon"   => ["density-density", "displacement-displacement"],
    "electron-phonon"   => ["density-density", "density-displacement"]
)

List of all Jastrow parameters that can be added, along with the corresponding ordering type.

Determinantal Parameters

VariationalMC.DeterminantalParameters — Type

DeterminantalParameters{S<:AbstractString, T<:Number}

Determinantal parameters allowing for different types of ordering. This include charge order in the form of "μ" and $"density"$, spin order in the form of "spin-x" and "spin-x", and pairing in the form of "s-wave" and "d-wave".

Fields

order_type::Vector{S}: The class of each order parameter, being "charge", "spin", or "pair".

param_name::Vector{S}: Specific order parameters names.

symmetry::Vector{S}: Symmetry of each order parameter, being "uniform" or "site-dependent".

p::Vector{Vector{T}}: Vector of order parameters.

qₚ::Vector{AbstractFloat}: Center-of-mass pairing momentum when introducing site-dependent pairing.

optimize::Vector{Bool}: Determines whether an order parameter will be optimized.

orbital::Vector{Int}: All orbital species where each order parameter applies.

VariationalMC.DeterminantalParameters — Method

DeterminantalParameters(;
    tight_binding_parameters::TightBindingParameters{T},
    model_geometry::ModelGeometry{D,T},
    rng::AbstractRNG
) where {D, T<:AbstractFloat}

Initialize and return an instance of `DeterminantalParameters'.

VariationalMC.add_parameter! — Function

add_parameter!(
    determinantal_parameters::DeterminantalParameters;
    param_name::Union{S, Nothing} = nothing,
    order_type::Union{S, Nothing} = nothing,
    symmetry::Union{S, Nothing} = nothing,
    optimize::Union{Bool, Nothing} = nothing,
    orbital::Union{Vector{Int}, Nothing} = nothing,
    p::Union{Vector{AbstractFloat}, Nothing} = nothing,
    model_geometry::Union{ModelGeometry, Nothing} = nothing,
    rng::Union{AbstractRNG, Nothing} = nothing
) where {S<:AbstractString}

Adds a parameters definition to determinantal_parameters.

This method first checks that the given parameter is not already defined. If it is, this method simply flags the parameter with that name for optimization. If the given parameter is not already defined, then it is added to the definitons in determinantal_parameters.

Jastrow Parameters

VariationalMC.JastrowParameters — Type

JastrowParameters{E<:AbstractFloat}

Jastrow pseudopotentials for different types of Jastrow factor. These include density-density, spin-spin, phonon-phonon, and electron-phonon.

Fields

particle_pair::AbstractString: Pair of particles being subject to Jastrow pseudopotentials, being "electron-electron", "phonon-phonon", and "electron-phonon".

order_pair::AbstractString: Pair of particle ordering types that are being reduced with Jastrow pseupotentials, being "density-density", "spin-spin", "density-displacement", and "displacement-displacement".

irr_indices::Vector{Int}: Set of (irreducible) single indices identifying equivalence classes of lattice site index pairs. Note that irreducible indices start at 0.

mean_v::Vector{E}: Strength of each pseudopotential for each irreducible index.

optimize::Bool: Optimization flag for whether the pseudopotentials will be optimized. If true, all parameters will be optimized excpet for ones associated with the largest irreducible index.

orbitals::Vector{Int}: Orbital IDs for each irreducible index.

VariationalMC.JastrowParameters — Method

JastrowParameters(;
    particle_pair::AbstractString,
    order_pair::AbstractString,
    orbitals::Vector{Int},
    optimize::Bool,
    model_geometry::ModelGeometry,
    rng::AbstractRNG,
    v::Union{Vector{E}, Nothing} = nothing
) where {E<:AbstractFloat}

Initialize and return an instance of `JastrowParameters'.

Monte Carlo

Variational Wavefunction

Metropolis Updater

Stochastic Reconfiguration

Variational Wavefunction

VariationalMC.DeterminantalWavefunction — Type

DeterminantalWavefunction{T<:Number, E<:AbstractMatrix{T}}

A mutable struct containing all the parameters needed to characterize the determinantal part of the trial wavefunction.

Fields

W::E: Equal-time Green's function matrix.

D::E: Slater matrix.

M::E: Reduced form of the unitary matrix U.

A::Vector{E}: Set of logarithmic derivative operator matrices for all parameters that will be optimized.

nq_updates_W::Int: Tracker for the number of quick updates that have been performed on the equal-time Green's function.

VariationalMC.DeterminantalWavefunction — Method

DeterminantalWavefunction(;
    # KEYWORD ARGUMENTS
    determinantal_parameters::DeterminantalParameters,
    hamiltonian::Hamiltonian{T,E},
    particle_configuration::ParticleConfiguration,
    model_geometry::ModelGeometry
) where {T, E}

Initialize and return an instance of DeterminantalWavefunction.

VariationalMC.AbstractJastrowFactor — Type

abstract type AbstractHST{T<:Number, R<:AbstractFloat} end

Abstract type to represent a Jastrow factor.

VariationalMC.FermionJastrowFactor — Type

FermionJastrowFactor{E<:AbstractFloat} <: AbstractJastrowFactor{E}

BosonJastrowFactor{E<:AbstractFloat} <: AbstractJastrowFactor{E}

MixedJastrowFactor{E<:AbstractFloat} <: AbstractJastrowFactor{E}

A mutable struct containing all the parameters needed to characterize a Jastrow factor involving fermions and/or bosons (phonons).

Fields

T::Vector{Vector{E}}: Vector of Jastrow factor elements.

orbitals::Vector{Int}: Orbital IDs subject to Jastrow pseudopotentials.

nq_updates_T::Int: Tracker for the numbner of quick updates that have been performed to vector T.

VariationalMC.JastrowFactor — Method

JastrowFactor(;
    # KEYWORD ARGUMENTS
    jastrow_parameters::JastrowParameters{E},
    particle_configuration::ParticleConfiguration,
    model_geometry::ModelGeometry
) where {E}

Initialize and return an instance of the JastrowFactor type depending upon the particle pair specified in jastrow_parameters.

Metropolis Updater

VariationalMC.local_fermion_update! — Function

local_fermion_update!(;
    # KEYWORD ARGUMENTS
    detwf::DeterminantalWavefunction{T},
    jastrow_factors::Union{Tuple{<:AbstractJastrowFactor{T}},Nothing} = nothing,
    tight_binding_parameters::TightBindingParameters,
    jastrows_parameters::Union{Tuple{JastrowParameters{T}},Nothing} = nothing,
    particle_configuration::ParticleConfiguration,
    model_geometry::ModelGeometry,
    δW::AbstractFloat,
    n_stab_W::Int,
    rng::AbstractRNG
) where {T}

Performs a local update to the fermionic configuration by attempting to move a particle with ID $\beta$ at a randomly selected lattice site $k$ to another random site $l$, with acceptance determined by the Metropolis-Hasting algorithm.

KEYWORD ARGUMENTS

detwf::DeterminantalWavefunction{T}: Current instance of DeterminantalWavefunction.

jastrow_factors::Union{Tuple{<:AbstractJastrowFactor{T}},Nothing} = nothing: Tuple of jastrow_factor.

tight_binding_parameters::TightBindingParameters: Instance of TightBindingParameters.

jastrows_parameters::Union{Tuple{JastrowParameters{T}},Nothing} = nothing: Tuple of jastrow_parameters.

particle_configuration::ParticleConfiguration: Instance of ParticleConfiguration.

model_geometry::ModelGeometry: Instance of ModelGeometry.

δW::AbstractFloat: Maximum allowed error in the equal-time Green's function matrix W.

δT::Union{AbstractFloat,Nothing} = nothing: Maximum allowed error in the Jastrow vector T.

n_stab_W::Int: Frequency of numerical stability checks performed on the matrix W.

n_stab_T::Union{Int,Nothing} = nothing: Frequency of numerical stability checks performed on the vector T.

rng::AbstractRNG: Random number generator.

Stochastic Reconfiguration

VariationalMC.Optimizer — Type

Optimizer{T<:Number}

A mutable struct containing all the parameters needed to characterize a Stochastic Reconfiguration optimizer.

Fields

var_params::Vector{T}: All variational parameters in the model.

η::AbstractFloat: Parameter which ensures the stability of the optimization.

dt::AbstractFloat: Parameter which controls the rate of optimization.

S::Matrix{T}: Covariance matrix.

f::Vector{T}: Forces.

VariationalMC.update_optimizer! — Function

update_optimizer!(;
    optimizer::Optimizer{T},
    bin_size::Int,
    determinantal_parameters::DeterminantalParameters{S,T},
    jas_parameters::Union{Tuple{JastrowParameters{T}}, Nothing} = nothing
) where {S, T<:Number}

Updates the optimizer by performing Stochastic Reconfiguration. The updated parameters are then pushed back to their respective containers.

NOTE

During the optimization step, this function normalizes all measurements so that the optimizer can construct S and f. Ensure that the opt field in the write_measurements! method is set to true.

Measurement Methods

Measurement Names

Initialize Measurements

Make Measurements

Write Measurements

Process Measurements

Export Measurements

Measurement Names

VariationalMC.GLOBAL_MEASUREMENTS — Constant

GLOBAL_MEASUREMENTS = (
    "density",
    "double_occ",
    "spin-z",
    "energy_per_site"
)

List of all the global measurements that are made.

VariationalMC.LOCAL_MEASUREMENTS — Constant

LOCAL_MEASUREMENTS = Base.ImmutableDict(
    "density"                  => "ORBITAL_ID",
    "site-dependent_density"   => "ORBITAL_ID",
    "double_occ"               => "ORBITAL_ID",
    "spin-z"                   => "ORBITAL_ID",
    "site-dependent_density"   => "ORBITAL_ID"
    "hopping_energy"           => "HOPPING_ID",
    "hubbard_energy"           => "HUBBARD_ID",
    "ext_hub_energy"           => "EXT_HUB_ID",
    "phonon_kin_energy"        => "PHONON_ID",
    "phonon_pot_energy"        => "PHONON_ID",
    "X"                        => "PHONON_ID",
    "holstein_energy"          => "HOLSTEIN_ID",
    "ssh_energy"               => "SSH_ID",
)

List of all the local measurements than can be made, with a mapping to the corresponding type of ID each measurement is reported in terms of.

VariationalMC.CORRELATION_FUNCTIONS — Constant

CORRELATION_FUNCTIONS = Base.ImmutableDict(
    "density"          => "ORBITAL_ID",
    "spin_x"           => "ORBITAL_ID",
    "spin_z"           => "ORBITAL_ID",
    "pair"             => "BOND_ID",
)

List of all the correlation functions that can be measured, along with the corresponding type of ID the correlation measurement is reported in terms of. Correlation functions are well defined in both position and momentum space.

Initialize Measurements

VariationalMC.initialize_measurement_container — Function

initialize_measurement_container(
    determinantal_parameters::DeterminantalParameters,
    model_geometry::ModelGeometry{D,T,N},
    N_opt::Int,
    N_sim::Int,
    opt_bin_size::Int,
    sim_bin_size::Int;
    jas_parameters::Union{Tuple{JastrowParameters},Nothing}=nothing
) where {T<:AbstractFloat, D, N}

Initialize and return a measurement container of type NamedTuple.

VariationalMC.initialize_measurements! — Function

initialize_measurements!(
    measurement_container::NamedTuple,
    hubbard_model::HubbardModel{T}
) where {T<:AbstractFloat}

Initialize Hubbard model related measurements.

Initialized Measurements

hubbard_energy: Refer to measure_hopping_energy.

initialize_measurements!(
    measurement_container::NamedTuple,
    extended_hubbard_model::ExtendedHubbardModel{T}
) where {T<:AbstractFloat}

Initialize Extended Hubbard model related measurements.

Initialized Measurements

ext_hubbard_energy: Refer to measure_ext_hubbard_energy.

VariationalMC.initialize_correlation_measurements! — Function

initialize_correlation_measurements!(;
    measurement_container::NamedTuple,
    correlation::String,
    pairs::AbstractVector{NTuple{2,Int}},
    time_displaced::Bool,
    integrated::Bool = false
)  where {T<:AbstractFloat, D, N}

Initialize measurements of correlation for all ID pairs; refer to CORRELATION_FUNCTIONS for ID type associated with each correlation measurement. The name correlation must therefore also appear in [CORRELATION_FUNCTIONS]@ref.

Make Measurements

VariationalMC.make_measurements! — Function

make_measurements!(
    measurement_container::NamedTuple;
    detwf::DeterminantalWavefunction{T},
    jas_factors::Union{Tuple{<:AbstractJastrowFactor{T}}, Nothing} = nothing,
    tight_binding_parameters::TightBindingParameters{T, E},
    jas_parameters::Union{Tuple{JastrowParameters{E}}, Nothing} = nothing,
    coupling_parameters::Tuple,
    particle_configuration::ParticleConfiguration,
    model_geometry::ModelGeometry;
    opt_step::Bool = false
) where {T<:Number, E<:AbstractFloat}

Make measurements. If used during the optimization phase of the simulation, set opt_step = true, which will activate the measurements of the logarithmic derivatives.

Write Measurements

VariationalMC.write_measurements! — Function

write_measurements!(;
    # KEYWORD ARGUMENTS
    measurement_container::NamedTuple,
    simulation_info::SimulationInfo,
    model_geometry::ModelGeometry,
    bin::Int,
    bin_size::Int,
    opt_step::Bool = false
)

Write the measurements contained in measurement_container to file. Measurements are written to file in HDF5 format.

VariationalMC.merge_bins — Function

merge_bins(
    simulation_info::SimulationInfo;
    opt::Bool = false
)

Merge the separate HDF5 files containing the binned measurements into a single HDF5 file. This is true even if the HDF5 "files" containing the binned data were held in memory during the simulation (simulation_info.write_bins_concurrent = false) instead of being actively written to file during the simulation (simulation_info.write_bins_concurrent = true). Set opt = true if merging binds from the optimization step.

VariationalMC.rm_bins — Function

rm_bins(
    comm::MPI.Comm,
    datafolder::String
)

rm_bins(
    datafolder::String
)

Delete the binned data stored in the datafolder directory. This function essentially deletes the directory datafolder/bins and its contents.

Process Measurements

VariationalMC.save_simulation_info — Function

save_simulation_info(
    # ARGUMENTS
    sim_info::SimulationInfo,
    metadata = nothing;
    # KEYWORD ARGUMENTS
    filename = @sprintf "simulation_info_sID-%d_pID-%d.toml" sim_info.sID sim_info.pID
)

Save the contents sim_info to a TOML file, and add an optional additional table to the output file based on the contents of a dictionary metadata.

VariationalMC.process_measurements — Function

process_measurements(
    # ARGUMENTS
    comm::MPI.Comm;
    # KEYWORD ARGUMENTS
    datafolder::String,
    ...
)

process_measurements(;
    # KEYWORD ARGUMENTS
    datafolder::String,
    ...
)

Process the HDF5 files containing the binned data generated by a DQMC simulation, and then write final statistics to a new HDF5 file(s). There is also functionality available that allows the final stats to then be exported to CSV files as well.

When the first passed arguments is comm::MPI.Comm, then the rebinning of data can be parallelized using MPI. However, it will only be parallelized if parallelize_rebinning = true.

Keyword Arguments

This function has many possible keyword arguments. Below I group them together into related categories and define their meanings.

Required Keyword Arguments

datafolder::String: Specify the directory generated by a DQMC simulation into which all the results were written.

Keyword for Controlling Workflow

parallelize_rebinning::Bool = true: If comm::MPI.Comm is passed as first argument to this function, then if this is true then rebinning is done on each MPI process and then collected together on the root process where jackknife is used to calculate mean and error of observables. Otherwise all the rebinning is performed by the root process alone.

n_bins::Union{Int, Nothing} = nothing: Number of bins used per walker to calculate statistics. If nothing then set equal to the number of data bins written to file during the simulation. Must be a factor of the number of bins written to file per walker during the DQMC simulation.

pIDs::Union{Int,Vector{Int}} = Int[]: Specifies for which process IDs to calculate average statistics. If pIDs = Int[], the calculate for all process IDs. If comm::MPI.Comm is passed as first function argument then pIDs must be of type Vector{Int} and not Int.

filename_prefix::String = "stats": Start of filename for HDF5 containing final statistics. HDF5 files containing statistics for a single process ID will end with pID-$(pID).h5.

rm_binned_data::Bool = false: Whether to delete the binned data after final statistics are computed.

Keywords for Exporting Statistics to CSV

export_to_csv::Bool = true: Whether to export the final statistics to CSV file.

scientific_notation::Bool = false: Whether to use scientific notation when exporting statistics to CSV file.

decimals::Int = 9: How many decimal places to include when exporting statistics to CSV files.

delimiter::String = " ": Delimiter used when writing CSV files.

Keyword Acting as Boolean Flags Indicating Which Statistics to Compute

process_global_measurements::Bool = true: Whether to calculate the statistics for global measurements.

process_local_measurements::Bool = true: Whether to calculate the statistics for local measurements.

process_all_equal_time_measurements::Bool = true: Whether to calculate statistics for all equal-time correlation measurements.

Export Measurements

VariationalMC.export_global_stats_to_csv — Function

export_global_stats_to_csv(
    # ARGUMENTS
    comm::MPI.Comm;
    # KEYWORD ARGUMENTS
    datafolder::String,
    h5filename::String = "stats.h5",,
    csvfilename_prefix::String = "global",
    measurements::Vector{String} = String[],
    decimals::Int = 6,
    scientific_notation::Bool = false,
    delimiter::String = " "
)

export_global_stats_to_csv(;
    # KEYWORD ARGUMENTS
    datafolder::String,
    h5filename::String = "stats.h5",,
    csvfilename_prefix::String = "global",
    measurements::Vector{String} = String[],
    decimals::Int = 6,
    scientific_notation::Bool = false,
    delimiter::String = " "
)

This function writes the global measurement statistics stored in the h5filename HDF5 file found in the directory datafolder to CSV file, returning the name of the CSV file that was written. The measurements keyword argument specifies the measurements to be exported. If measurements = String[], then all measurements are exported.

VariationalMC.export_local_stats_to_csv — Function

export_local_stats_to_csv(
    # ARGUMENTS
    comm::MPI.Comm;
    # KEYWORD ARGUMENTS
    datafolder::String,
    h5filename::String = "stats.h5",
    csvfilename_prefix::String = "local",
    measurements::Vector{String} = String[],
    decimals::Int = 6,
    scientific_notation::Bool = false,
    delimiter::String = " "
)

export_local_stats_to_csv(;
    # KEYWORD ARGUMENTS
    datafolder::String,
    h5filename::String = "stats.h5",
    csvfilename_prefix::String = "local",
    measurements::Vector{String} = String[],
    decimals::Int = 6,
    scientific_notation::Bool = false,
    delimiter::String = " "
)

This function writes the local measurement statistics stored in the h5filename HDF5 file found in the directory datafolder to CSV file, returning the name of the CSV file that was written. The measurements keyword argument specifies the measurements to be exported. If measurements = String[], then all measurements are exported.

VariationalMC.export_correlation_stats_to_csv — Function

export_correlation_stats_to_csv(
    # ARGUMENTS
    comm::MPI.Comm;
    # KEYWORD ARGUMENTS
    datafolder::String,
    correlation::String,
    type::String,
    space::String,
    h5filename::HDF5.File = "stats.h5",
    decimals::Int = 6,
    scientific_notation::Bool = false,
    delimiter::String = " "
)

export_correlation_stats_to_csv(;
    # KEYWORD ARGUMENTS
    datafolder::String,
    correlation::String,
    type::String,
    space::String,
    h5filename::HDF5.File = "stats.h5",
    decimals::Int = 6,
    scientific_notation::Bool = false,
    delimiter::String = " "
)

Export statistics for specified type of correlation function from HDF5 file to a CSV file.