API

ModelGeometry Type and Methods

ModelGeometry

VariationalMC.ModelGeometry — Type

ModelGeometry( unit_cell::UnitCell, 
               lattice::Lattice, 
               bond::Vector{Vector{Any}} )

A type defining model geometry.

Parameters Types and Methods

TightBindingModel

SpinModel

DeterminantalParameters

JastrowParameters

VariationalMC.TightBindingModel — Type

TightBindingModel( t₀::Float64,
                   t₁::Float64 )

A type defining a non-interacting tight binding model.

VariationalMC.SpinModel — Type

SpinModel( J₁::Float64,
           J₂::Float64,
           J₃::Float64 )

A type defining a spin model.

VariationalMC.DeterminantalParameters — Type

DeterminantalParameters( pars::Vector{AbstractString}, 
                         vals::Vector{AbstractFloat}, 
                         num_detpars::Int )

A type defining a set of variational parameters for the determinantal wavefunction.

VariationalMC.JastrowParameters — Type

JastrowParameters( jastrow_type::String,
                   jpar_map::OrderedDict{Any, Any},
                   num_jpars::Int,
                   num_jpar_opts::Int )

A type defining quantities related to Jastrow variational parameters.

DeterminantalWavefunction Type and Methods

DeterminantalWavefunction

get_determinantal_wavefunction

VariationalMC.DeterminantalWavefunction — Type

DeterminantalWavefunction( W::Matrix{ComplexF64}, 
                           D::Matrix{ComplexF64}, 
                           M::Matrix{ComplexF64}
                           U_int::Matrix{ComplexF64}, 
                           A::Vector{Any}, 
                           ε::Vector{Float64}, 
                           pconfig::Vector{Int64} )

A type defining quantities related to a determinantal wavefunction.

VariationalMC.get_determinantal_wavefunction — Function

get_determinantal_wavefunction( tight_binding_model::TightBindingModel, 
                                determinantal_parameters::DeterminantalParameters, 
                                Ne::Int64, 
                                nup::Int64, 
                                ndn::Int64,
                                model_geometry::ModelGeometry, 
                                rng::Xoshiro)::DeterminantalWavefunction

Constructs a variational wavefunction based on parameters given by the tight binding model and determinantal parameters. Returns an instances of the DeterminantalWavefunction type. If no particle configuration is specified, a random configuration will be generated.

tight_binding_model::TightBindingModel: parameters for a non-interacting tight-binding model.

determinantal_parameters::DeterminantalParameters: set of determinantal variational parameters.

optimize::NamedTuple: field of optimization flags.

Ne::Int: total number of electrons.

nup::Int: number of spin-up electrons.

ndn::Int: number of spin-down electrons.

model_geometry::ModelGeometry: contains unit cell and lattice quantities.

rng::Xoshiro: random number.

pconfig::Union{Nothing, Vector{Int}}=nothing: current particle configuration.

Jastrow Types and Methods

JastrowFactor

get_jastrow_factor

VariationalMC.JastrowFactor — Type

JastrowFactor( Tvec_f::Vector{Float64}, 
               Tvec_b::Vector{Float64} )

A type defining quantities related to a Jastrow factor.

VariationalMC.get_jastrow_factor — Function

get_jastrow_factor( jastrow_parameters::JastrowParameters, 
                    detwf::DeterminantalWavefunction, 
                    model_geometry::ModelGeometry, 
                    pht::Bool )::JastrowFactor

Constructs specified Jastrow factor and returns a instance of the JastrowFactor type.

jastrow_parameters::JastrowParameters: current set of Jastrow parameters.

detwf::DeterminantalWavefunction: current variational wavefunction.

model_geometry::ModelGeometry: contains unit cell and lattice quantities.

pht::Bool: whether model is particle-hole transformed.

Markov Methods

local_fermion_update!

VariationalMC.local_fermion_update! — Function

local_fermion_update!( detwf::DeterminantalWavefunction,
                       Ne::Int, 
                       model_geometry::ModelGeometry, 
                       pht::Bool, 
                       δW::Float64, 
                       δT::Float64, 
                       rng::Xoshiro )

Performs a local update to the electronic sector, with a certain number of equilibration steps.

detwf::DeterminantalWavefunction: current determinantal variational wavefunction.

Ne::Int: total number of electrons.

model_geometry::ModelGeometry: contains unit cell and lattice quantities.

n_stab_W: frequency of Green's function stabilization steps.

δW::Float64: error threshold for the Green's function.

rng::Xoshiro: random number generator.

local_fermion_update!( detwf::DeterminantalWavefunction, 
                       jastrow_factor::JastrowFactor, 
                       jastrow_parameters::JastrowParameters,
                       Ne::Int, 
                       model_geometry::ModelGeometry, 
                       pht::Bool, 
                       δW::Float64, 
                       δT::Float64, 
                       rng::Xoshiro )

Performs a local update to the electronic sector, with a certain number of equilibration steps.

detwf::DeterminantalWavefunction: current determinantal variational wavefunction.

jastrow_parameters::JastrowParameters: current set of Jastrow variational parameters.

jastrow::Jastrow: current Jastrow factor.

Ne::Int: total number of electrons.

model_geometry::ModelGeometry: contains unit cell and lattice quantities.

n_stab_W: frequency of Green's function stabilization steps.

n_stab_T: frequency of T vector stabilization steps.

δW::Float64: error threshold for the Green's function.

rng::Xoshiro: random number generator.

ParticleConfiguration Types and Methods

get_particle_numbers

get_particle_density

VariationalMC.get_particle_numbers — Function

get_particle_numbers( density::Float64 )::NTuple{4, Int64}

Given a particle density, returns the total number of particles Np, number of spin-up particles Npu, number of spin-down particles Npd, and total number of electrons Ne.

density::Float64: desired electronic density.

VariationalMC.get_particle_density — Function

get_particle_density( nup::Int64, 
                      ndn::Int64 )::Tuple{Float64, Int64, Int64}

Given the number of spin-up electrons nup, and number of spin-down electrons ndn, returns the particle density, total number of particles Np, and the total number of electrons Ne.

nup::Int: number of spin-up electrons.

ndn::Int: number of spin-down electrons.

Optimizer Methods

stochastic_reconfiguration!

VariationalMC.stochastic_reconfiguration! — Function

stochastic_reconfiguration!( measurement_container::NamedTuple, 
                             determinantal_parameters::DeterminantalParameters, 
                             η::Float64, 
                             dt::Float64, 
                             bin::Int,
                             bin_size::Int64 )::Nothing

Updates variational parameters through stochastic optimization.

measurement_container::NamedTuple: container where measurements are stored.

determinantal_parameters::DeterminantalParameters: set of determinantal variational parameters.

η::Float64: optimization stabilization factor.

dt::Float64: optimization rate.

bin::Int: current bin number.

bin_size::Int64: length of the current bin.

stochastic_reconfiguration!( measurement_container::NamedTuple, 
                             determinantal_parameters::DeterminantalParameters, 
                             jastrow_parameters::JastrowParameters, 
                             η::Float64, 
                             dt::Float64, 
                             bin::Int, 
                             bin_size::Int64 )::Nothing

Updates variational parameters through stochastic optimization.

measurement_container::NamedTuple: container where measurements are stored.

determinantal_parameters::DeterminantalParameters: set of determinantal variational parameters.

jastrow_parameters::JastrowParameters: set of Jastrow variational parameters.

η::Float64: optimization stabilization factor.

dt::Float64: optimization rate.

bin::Int: current bin number.

bin_size::Int64: length of the current bin.

Measurement Methods

Intitialize Measurements

initialize_measurement_container

initialize_measurement_directories

VariationalMC.initialize_measurement_container — Function

initialize_measurement_container( N_opts::Int, 
                                  opt_bin_size::Int, 
                                  N_bins::Int, 
                                  bin_size::Int,
                                  determinantal_parameters::DeterminantalParameters, 
                                  model_geometry::ModelGeometry )::NamedTuple

Initializes a set of dictionaries containing generic arrays for storing measurements. Each dictionary in the container has [keys => values] of the form: [observable name => (local value(s), current bin value(s))]

N_opts::Int: number of optimization updates.

opt_bin_size::Int: length of an optimization bin.

N_bins::Int: number of simulation bins.

bin_size::Int: length of a simulation bin.

determinantal_parameters::DeterminantalParameters: set of determinantal variational parameters.

model_geometry::ModelGeometry: contains unit cell and lattice quantities.

initialize_measurement_container( N_opts::Int, 
                                  opt_bin_size::Int, 
                                  N_bins::Int, 
                                  bin_size::Int,
                                  determinantal_parameters::DeterminantalParameters, 
                                  jastrow_parameters::JastrowParameters,
                                  model_geometry::ModelGeometry )::NamedTuple

Initializes a set of dictionaries containing generic arrays for storing measurements. Each dictionary in the container has [keys => values] of the form: [observable name => (local value(s), current bin value(s))]

N_opts::Int: number of optimization updates.

opt_bin_size::Int: length of an optimization bin.

N_bins::Int: number of simulation bins.

bin_size::Int: length of a simulation bin.

determinantal_parameters::DeterminantalParameters: set of determinantal variational parameters.

jastrow_parameters::JastrowParameters: set of Jastrow variational parameters.

model_geometry::ModelGeometry: contains unit cell and lattice quantities.

VariationalMC.initialize_measurement_directories — Function

initialize_measurement_directories( simulation_info::SimulationInfo, 
                                    measurement_container::NamedTuple )::Nothing

Creates file directories and subdirectories for storing measurements.

simulation_info::SimulationInfo: contains datafolder information.

measurement_container::NamedTuple: container where measurements are contained.

Make Measurements

make_measurements!

VariationalMC.make_measurements! — Function

make_measurements!( measurement_container::NamedTuple, 
                    detwf::DeterminantalWavefunction, 
                    tight_binding_model::TightBindingModel,
                    determinantal_parameters::DeterminantalParameters,
                    optimize::NamedTuple, 
                    model_geometry::ModelGeometry, 
                    Ne::Int64, 
                    pht::Bool )::Nothing

Measures global, optimization, and simulation observables.

measurement_container::NamedTuple: container where measurements are stored.

detwf::DeterminantalWavefunction: current determinantal wavefunction.

tight_binding_model::TightBindingModel: non-interacting tight-binding model.

determinantal_parameters::DeterminantalParameters: set of determinantal variational parameters.

model_geometry::ModelGeometry: contains unit cell and lattice quantities.

Ne::Int64: number of electrons.

pht::Bool: whether or not model is particle-hole transformed.

make_measurements!( measurement_container::NamedTuple, 
                    detwf::DeterminantalWavefunction, 
                    tight_binding_model::TightBindingModel,
                    determinantal_parameters::DeterminantalParameters,
                    jastrow_parameters::JastrowParameters,
                    jastrow_factor::JastrowFactor, 
                    model_geometry::ModelGeometry, 
                    Ne::Int64, 
                    pht::Bool )::Nothing

Measures global, optimization, and simulation observables.

measurement_container::NamedTuple: container where measurements are stored.

detwf::DeterminantalWavefunction: current determinantal wavefunction.

tight_binding_model::TightBindingModel: non-interacting tight-binding model.

determinantal_parameters::DeterminantalParameters: set of determinantal variational parameters.

jastrow_parameters::JastrowParameters: current set of Jastrow variational parameters.

jastrow_factor::JastrowFactor: current Jastrow factor.

model_geometry::ModelGeometry: contains unit cell and lattice quantities.

Ne::Int64: number of electrons.

pht::Bool: whether or not model is particle-hole transformed.

make_measurements!( measurement_container::NamedTuple, 
                    detwf::DeterminantalWavefunction, 
                    tight_binding_model::TightBindingModel, 
                    model_geometry::ModelGeometry, 
                    Ne::Int64, 
                    pht::Bool )

Measures simulation observables.

measurement_container::NamedTuple: container where measurements are stored.

detwf::DeterminantalWavefunction: current determinantal wavefunction.

tight_binding_model::TightBindingModel: non-interacting tight-binding model.

model_geometry::ModelGeometry: contains unit cell and lattice quantities.

Ne::Int64: number of electrons.

pht::Bool: whether or not model is particle-hoel transformed.

make_measurements!( measurement_container::NamedTuple, 
                    detwf::DeterminantalWavefunction, 
                    tight_binding_model::TightBindingModel, 
                    jastrow_parameters::JastrowParameters,
                    jastrow_factor::JastrowFactor, 
                    model_geometry::ModelGeometry, 
                    Ne::Int64, 
                    pht::Bool )

Measures simulation observables.

measurement_container::NamedTuple: container where measurements are stored.

detwf::DeterminantalWavefunction: current determinantal wavefunction.

tight_binding_model::TightBindingModel: non-interacting tight-binding model.

jastrow_parameters::JastrowParameters: current set of Jastrow variational parameters.

jastrow_factor::JastrowFactor: current Jastrow factor.

model_geometry::ModelGeometry: contains unit cell and lattice quantities.

Ne::Int64: number of electrons.

pht::Bool: whether or not model is particle-hole transformed.

Write Measurements

write_measurements!

VariationalMC.write_measurements! — Function

write_measurements!( measurement_container::NamedTuple, 
                     energy_bin::Vector{Any}, 
                     dblocc_bin::Vector{Any}, 
                     param_bin::Vector{Any} )::Nothing

DEBUG version of the write_measurements!() method. Will write binned energies, double occupancy, and parameters to specified vectors.

measurement_container::NamedTuple: container where measurements are stored.

energy_bin::Vector{Float64}: externally specified vector for storing energy measurements.

dblocc_bin::Vector{Float64}: externally specified vector for storing double occupancy measurements.

param_bin::Vector{Any}: externally specified vector for storing parameter measurements.

Process Measurements